requirements.txt
bcrypt==4.0.1
bleach==6.1.0
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
python-dotenv==1.0.0
six==1.16.0
webencodings==0.5.1
Werkzeug==2.3.0
WTForms==3.1.0